+++ smtlib-backends-z3/CHANGELOG.md
@@ -10,6 +10,7 @@ file.
 - add tests for documenting edge cases of the backends
   - what happens when sending an empty command
   - what happens when sending a command not producing any output
+- **(breaking change)** removed `Data.Default` instance of `Config`
 
 ### Changed
 - make test-suite compatible with `smtlib-backends-0.3`
@@ -17,6 +18,7 @@ file.
   `Z3.Config` object as argument, which one may use to set some solver options
   at initialization time
   - add corresponding examples in the test-suite
+- dropped dependency on `inline-c`
 
 ## v0.2 _(2022-12-16)_